requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pandas
numpy
tensorflow
joblib
python-dotenv